Initial commit of the Table class and related functionality and specs
The Table class is the meat of Akurum. It provides the ability to do queries against the database for rows based on index ids among other functionality. This is just a first pass at making this code work outside its original context and it needs a lot more work yet. But it does work, and the specs for basic functioanlity all work just fine.

Also added a Context derived class that includes some rspec matchers to make testing the Table class easier.